+++ NEWS.md
@@ -1,4 +1,11 @@
 # argonDash 0.2.0.9000
+
+## Major Note 
+- add support for shinyapps.io, RStudio connect and shiny server pro
+
+## Documentation
+- Better documentation of `argonDashPage()`
+
 ## Minor improvement
 - fix #3 thanks @noahbishop: when vertical is FALSE in the `argonDashSidebar`,
 navbar elements are horizontally displayed, as expected