+++ NEWS.md
@@ -7,6 +7,7 @@
 - Better documentation of `argonDashPage()`
 
 ## Minor improvement
+- add min-height to body
 - add `argonTable()`, `argonTableItems` and `argonTableItem`
 - add `argonAvatars()` and `argonAvatar` 
 - Fix #12: reduce vertical space between brand logo and menu items in `argonDashSidebar()`. Thanks @ShixiangWang