+++ NEWS.md
@@ -17,6 +17,7 @@
 navbar elements are horizontally displayed, as expected
 
 ## Bug fix
+- fix #21: icon_background has a default value in `argonInfoCard()`. Thanks @tusharmajhi9
 - fix #2: `argonDropdown()` correctly render in the navbar
 
 ## Breaking changes