+++ NEWS.md
@@ -18,6 +18,7 @@ navbar elements are horizontally displayed, as expected
 
 ## Bug fix
 - fix #21: icon_background has a default value in `argonInfoCard()`. Thanks @tusharmajhi9
+- fix #20: Remove hardcoded stat parameter, leave it user defined. Thanks @mkaranja
 - fix #2: `argonDropdown()` correctly render in the navbar
 
 ## Breaking changes